result is printed with type cast

+++ week-03/day-2/week3_day2_e01/main.c
@@ -9,7 +9,7 @@ main()
 {
 	int  i = 2;
 	float f = 13.5;
-	float result = f / i;
+	float result = f / (float)i;
 
 
     printf("%f" , result);
